Update models workflow create memory with text_source field instead of source, to prevent conflict with current memory model
