+++ backend/models/integrations.py
@@ -1,8 +1,4 @@
 from pydantic import BaseModel
-from datetime import datetime
-from enum import Enum
-from typing import Optional
-from models.memory import Geolocation
 
 
 class MemoryTimestampRange(BaseModel):
@@ -17,15 +13,5 @@ class ScreenPipeCreateMemory(BaseModel):
     timestamp_range: MemoryTimestampRange
 
 
-class WorkflowMemorySource(str, Enum):
-    audio = 'audio_transcript'
-    other = 'other_text'
-
-
-class WorkflowCreateMemory(BaseModel):
-    source: WorkflowMemorySource = WorkflowMemorySource.audio
-    started_at: Optional[datetime]
-    finished_at: Optional[datetime]
-    text: str
-    language: Optional[str] = None
-    geolocation: Optional[Geolocation] = None
+class EmptyResponse(BaseModel):
+    pass
